fortune-sheet/ai/requirements.txt
requests>=2.28.0
google-generativeai>=0.3.0
flask>=2.3.0
flask-cors>=4.0.0
python-dotenv>=1.0.0

